requirements.txt
Click==8.1.8
clint==0.5.1
pytest==7.2.1
pytest-cov==6.0.0
python-dotenv==1.0.0
pytube==12.1.2
requests==2.32.3
setuptools==45.2.0
static_ffmpeg==2.3
ffmpeg-python==0.2.0
yt-dlp>=2025.2.19
deepgram-sdk==2.4.0
boto3==1.26.143
ruff>=0.4.0
feedparser==6.0.10
PyYAML==6.0.1
librosa==0.10.2.post1
fastapi==0.111.0
PyJWT==2.9.0
cryptography==43.0.1
psutil==7.0.0
openai==1.98.0
pytest-asyncio==0.21.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
google-generativeai>=0.3.2
google-api-python-client>=2.100.0